donation

+++ categories/views.py
@@ -23,29 +23,35 @@ def product(request):
     return redirect('LOGIN')
 
 def productView(request, pk):
-    obj=item.objects.get(id=pk)
-    qty=1
-    if 'psub' in request.POST:
-        if '+' in request.POST.get('psub'):
-            qty=int(request.POST['qty'])+1
-        else:
-            qty=int(request.POST['qty'])
-            if qty>1:
-                qty=qty-1
-    return(render(request,'showProduct1.html', {'object':obj,'qty':qty}))
+    if 'email' in request.session:
+        obj=item.objects.get(id=pk)
+        qty=1
+        if 'psub' in request.POST:
+            if '+' in request.POST.get('psub'):
+                qty=int(request.POST['qty'])+1
+            else:
+                qty=int(request.POST['qty'])
+                if qty>1:
+                    qty=qty-1
+        return(render(request,'showProduct1.html', {'object':obj,'qty':qty}))
+    return redirect('LOGIN')
 
 def productUpdate(request, pk):
-    prod= get_object_or_404(item, pk=pk)
-    form = ProForm1(request.POST or None, instance=prod)
-    if form.is_valid():
-        form.save()
-        return redirect('PRODUCT')
-    return render(request, 'editProduct1.html', {'form':form})
+    if 'email' in request.session:
+        prod= get_object_or_404(item, pk=pk)
+        form = ProForm1(request.POST or None, instance=prod)
+        if form.is_valid():
+            form.save()
+            return redirect('PRODUCT')
+        return render(request, 'editProduct1.html', {'form':form})
+    return redirect('LOGIN')
 
 def productDelete(request, pk):
-    prod= get_object_or_404(item, pk=pk)   
-    prod.delete() 
-    return redirect('PRODUCT')
+    if 'email' in request.session:
+        prod= get_object_or_404(item, pk=pk)   
+        prod.delete() 
+        return redirect('PRODUCT')
+    return redirect('LOGIN')
 
 def add_to_cart(request, d, s1):
     if 'email' in request.session:
@@ -77,6 +83,7 @@ def show_mycart(request):
             p=p+i.book.price*i.quantity
         k=dict(zip(l,q))
         return(render(request,'mycartpage.html',{'k':k,'n':obj,'p':p}))
+    return redirect('LOGIN')
 
 def removecart(request,d):
     if 'email' in request.session:
@@ -84,55 +91,91 @@ def removecart(request,d):
         y=get_object_or_404(MyCart,book=d,person=o.pk)
         y.delete()
         return(redirect('MYCART'))
+    return redirect('LOGIN')
 
 def cartorder(request):
-    o=signUp.objects.get(email=request.session['email'])
-    obj=MyCart.objects.filter(person=o.pk)
-    l=[]
-    q=[]
-    p=0
-    s=''
-    for i in obj:
-        l.append(i.book)
-        q.append(i)
-        s+=i.book.title+" Id="+str(i.book.id)+" Quantity ="+str(i.quantity)+','
-        p=p+i.book.price*i.quantity
+    if 'email' in request.session:
+        o=signUp.objects.get(email=request.session['email'])
+        obj=MyCart.objects.filter(person=o.pk)
+        l=[]
+        q=[]
+        p=0
+        s=''
+        for i in obj:
+            l.append(i.book)
+            q.append(i)
+            s+=i.book.title+" Id="+str(i.book.id)+" Quantity ="+str(i.quantity)+','
+            p=p+i.book.price*i.quantity
 
-    if request.POST:
-        n=request.POST['name']
-        st=request.POST['state']
-        ct=request.POST['city']
-        ad=request.POST['address']
-        pin=request.POST['pincode']
-        ph=request.POST['phone']
-        dat=request.POST['date']
+        if request.POST:
+            n=request.POST['name']
+            st=request.POST['state']
+            ct=request.POST['city']
+            ad=request.POST['address']
+            pin=request.POST['pincode']
+            ph=request.POST['phone']
+            dat=request.POST['date']
+
+            iv=Order()
+            iv.oemail=request.session['email']
+            iv.services=s
+            iv.name=n
+            iv.service_date=dat
+            iv.contact=ph
+            iv.amount=p
+            iv.adddress=str(ad)+str(ct)+str(st)+'\n'+str(pin)
+
+            amount= p
+            print("Amount is ", amount)
+            print(type(amount),type(p))
+
+            client = razorpay.Client(auth=("rzp_test_ov7fBmU4EJwsAn", "AvmwLmd018H6gOgQrdeJPntX"))
+
+            payment = client.order.create({
+                'amount': amount*100,
+                'currency': 'INR',
+                'payment_capture': '1'
+            })
+
+            iv.save()
+            return(redirect('PAYMENT'))
+
+        k=dict(zip(l,q))
+        return(render(request,'orderpage.html',{'k':k,'p':p}))
     
-        iv=Order()
-        iv.oemail=request.session['email']
-        iv.services=s
-        iv.name=n
-        iv.service_date=dat
-        iv.contact=ph
-        iv.amount=p
-        iv.adddress=str(ad)+str(ct)+str(st)+'\n'+str(pin)
-        
-        amount= p
-        print("Amount is ", amount)
-        print(type(amount),type(p))
-        
-        client = razorpay.Client(auth=("rzp_test_ov7fBmU4EJwsAn", "AvmwLmd018H6gOgQrdeJPntX"))
-
-        payment = client.order.create({
-            'amount': amount*100,
-            'currency': 'INR',
-            'payment_capture': '1'
-        })
-
-        iv.save()
-        return(redirect('PAYMENT'))
-    
-    k=dict(zip(l,q))
-    return(render(request,'orderpage.html',{'k':k,'p':p}))
+    return redirect('LOGIN')
 
 def payment(request):
-    return(HttpResponse('Payment Successfully Done'))
+    if 'email' in request.session:
+        return(HttpResponse('Payment Successfully Done'))
+    return redirect('LOGIN')
+
+def donateMoney10(request):
+    return render(request,'donatemoney.html')
+        
+def donateMoney(request):
+    if 'email' in request.session:
+
+        user=signUp.objects.get(email=request.session['email'])
+        model=DonateMoney.objects.create(
+            person=user,
+        )
+        model.amount = request.POST.get('amount', 0)
+        model.save()
+        return render(request,'donatemoney.html',{'model':model})
+        
+    return redirect('LOGIN')
+    
+def donateMoney12(request):
+    if 'email' in request.session:
+
+        if request.method == "POST":
+            name = request.POST.get('amount')
+            amount = 50000
+
+            client = razorpay.Client(auth=("rzp_test_ov7fBmU4EJwsAn", "AvmwLmd018H6gOgQrdeJPntX"))
+
+            payment = client.order.create({'amount': amount, 'currency': 'INR', 'payment_capture': '1'})
+        return render(request, 'donatemoney.html')
+        
+    return redirect('LOGIN')
